move component specific hook to compoments file
