+++ frontend/src/components/OptionsMenu.tsx
@@ -6,7 +6,6 @@ import { useState, MouseEvent, Fragment, useEffect } from 'react'
 import HideColumnsDialog from './HideColumnsDialog'
 import DarkModeDialog from './DarkModeDialog'
 import PlanInfoDialog from './PlanInfoDialog'
-import useTooltip from './OptionsMenu_useTooltip'
 
 const options = [
   {
@@ -111,4 +110,39 @@ export default function OptionsMenu() {
       })}
     </Fragment>
   )
+}
+
+function useTooltip() {
+  const [tooltipOpen, setTooltipOpen] = useState(false)
+
+  useEffect(() => {
+    let timeout: number | undefined
+    if (localStorage.getItem('hideColumnsTooltip') === null) {
+      const handleAnyClick = () => {
+        localStorage.setItem('hideColumnsTooltip', 'true')
+        setTooltipOpen(false)
+        clearTimeout(timeout)
+      }
+
+      setTooltipOpen(true)
+
+      timeout = setTimeout(() => {
+        localStorage.setItem('hideColumnsTooltip', 'true')
+        setTooltipOpen(false)
+        document.removeEventListener('click', handleAnyClick)
+        document.removeEventListener('touchstart', handleAnyClick)
+      }, 5000)
+
+      document.addEventListener('click', handleAnyClick)
+      document.addEventListener('touchstart', handleAnyClick)
+
+      return () => {
+        clearTimeout(timeout)
+        document.removeEventListener('click', handleAnyClick)
+        document.removeEventListener('touchstart', handleAnyClick)
+      }
+    }
+  }, [])
+
+  return tooltipOpen
 }
